// graphql-types.ts
export type Language =
  | 'EN'
  | 'RU';
// ...
export type BioQueryQuery = { site?: { siteMetadata?: { i18n?: Array<{ lang?: Language | null, siteMetadata?: { author?: { name?: string | null, summary?: string | null } | null } | null } | null> | null } | null } | null };